+++ generators/app/templates/app-electron-vuetify/src/plugins/vuetify.ts
@@ -1,0 +1,8 @@
+import Vue from 'vue';
+// @ts-ignore
+import Vuetify from 'vuetify/lib';
+import 'vuetify/src/stylus/app.styl';
+
+Vue.use(Vuetify, {
+  iconfont: 'md',
+});